fix bug when user can not be deleted if he hasn't got image url

+++ backend/src/main/java/com/book/exchange/backend/user/UserService.java
@@ -193,7 +193,8 @@ public class UserService {
 
     public void deleteUserBySlug(String slug) {
         UserEntity user = getUserBySlug(slug);
-        imageService.deleteImageBySlug(user.getImageUrl().substring(imageUrl.length()));
+        if (user.getImageUrl() != null)
+            imageService.deleteImageBySlug(user.getImageUrl().substring(imageUrl.length()));
         userRepository.delete(user);
     }
